Changed again cave spider spawn (Fixed #9)
Was still crashing because it could try to find a nextInt with parameter 0. Totally changed the way we compute the randomness: now only spawn below 64 y-level (if bottom of the world is negative, add it to 64. so in 1.18 it means below 0 y-level) and the deeper we are, the bigger the chance to spawn is