Added custom cobweb block #19
A cobweb projectile will spawn a messy cobweb ob entity hit. This one is like a normal cobweb, but a bit more fragile (breaks a bit more quickly), despawn in 5 to 10s, don't loot strings, and can't be placed on fluids. Due to this fragility, the projectile is now also discard if it touches water