Zombies have attributes variations and updated base values
Zombies knockback resistance is updated and randomized. It can be
greatly increased if it spawns as an "unstoppable" zombie.
Zombies spawn reinforcement is increased quite a lot and randomized. It
can be greatly increased if it spawns as a "caller" zombie.
"tank" and "runner" zombies variants are also added, with increased
health and speed respectively.
These zombie variants can't be applied to baby zombies.
Brainless, leader and siege are replaced by these new settings.