Add zombie horse trap
Add a zombie horse trap, similar to the skeleton horse trap. Lightning can summon a zombie horse trap. If triggered by a player, it summons three other zombie horses ridden by zombies (similarly to the skeleton horse trap).
Every zombie is adult. One is a regular zombie with an iron sword, one is a husk with an iron shovel, one is a drowned with a trident, and the last one is a zombified piglin with a golden sword. They all spawn with an iron helmet, just like skeletons from the trap.